+++ src/main/java/br/gov/servicos/cadastro/LocalidadeRepositorio.java
@@ -13,7 +13,7 @@ public class LocalidadeRepositorio{
 	@PersistenceContext
 	private EntityManager entity;
 	
-	public boolean existeLocalidade(Long idLocalidade) {
+	public boolean existeLocalidade(Integer idLocalidade) {
 		StringBuilder sql = new StringBuilder();
 		sql.append("select count (lo) from Localidade lo ")
 			.append("where lo.id = :idLocalidade");
@@ -25,7 +25,7 @@ public class LocalidadeRepositorio{
 		return (qtd > 0) ? true : false; 
 	}
 	
-	public Localidade find(Long id){
+	public Localidade find(Integer id){
 		try{
 			return entity.find(Localidade.class, id);
 		} catch (NoResultException e) {
